+++ tools/src/main/scala/com/nvidia/spark/rapids/tool/qualification/RunningQualificationApp.scala
@@ -93,7 +93,7 @@ class RunningQualificationApp(readScorePercent: Int = QualificationArgs.DEFAULT_
           QualOutputWriter.getSummaryHeaderStringsAndSizes(Seq(info), appIdMaxSize)
         val headerStr = QualOutputWriter.constructOutputRowFromMap(headersAndSizes,
           delimiter, prettyPrint)
-        val sumsToWrite = QualOutputWriter.calculateEstimatedInfoSummary(Seq(info))
+        val sumsToWrite = QualificationAppInfo.calculateEstimatedInfoSummary(Seq(info))
         val appInfoStr = QualOutputWriter.constructAppSummaryInfo(sumsToWrite.head,
           headersAndSizes, appIdMaxSize,
           delimiter, prettyPrint)
